Started working on a save function/Cleaned up a bit
I cleaned a few repetitive functions and I started working on a decoder.
I really couldn't figure out how to turn the shit to binary because I
felt like it, but turning them all into chars looks nice, I'm going to
have it where you need to save it all in a certain order for an object,
and load it in a certain order for an object. I'll see how it goes and
maybe change it if it doesn't work efficiently!